+++ CHANGELOG.md
@@ -7,6 +7,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- `GarbageCollector` pool wrapper for retaining only cached acceleration structures, buffers, and
+  images that support recently observed requests.
+
 ## [0.14.5] - 2026-07-13
 
 ### Added